+++ src/i18next.translate.js
@@ -199,7 +199,7 @@ function _find(key, options){
         optionWithoutCount.defaultValue = o.pluralNotFound;
 
         var pluralKey = ns + o.nsseparator + key + o.pluralSuffix;
-        var pluralExtension = pluralExtensions.get(currentLng, options.count);
+        var pluralExtension = pluralExtensions.get(lngs[0], options.count);
         if (pluralExtension >= 0) {
             pluralKey = pluralKey + '_' + pluralExtension;
         } else if (pluralExtension === 1) {